feat: 优化资产导入, 可以添加节点全称，并自动创建
